+++ BE/src/main/java/com/codesquad/baseball06/dto/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.codesquad.baseball06.model;
+package com.codesquad.baseball06.dto;
 
 public class ApiResponse<T> {
