+++ BE/src/main/java/com/codesquad/baseball06/model/dao/PlayerDao.java
@@ -1,16 +1,12 @@
-package com.codesquad.baseball06.dao;
+package com.codesquad.baseball06.model.dao;
 
-import com.codesquad.baseball06.dao.mapper.BatterReturnDtoMapper;
-import com.codesquad.baseball06.dao.mapper.PitcherReturnDtoMapper;
 import com.codesquad.baseball06.dto.BatterReturnDto;
 import com.codesquad.baseball06.dto.PitcherReturnDto;
-import com.codesquad.baseball06.model.Batter;
-import com.codesquad.baseball06.model.Pitcher;
-import com.codesquad.baseball06.model.Player;
-import com.codesquad.baseball06.model.Team;
-import java.util.Optional;
+import com.codesquad.baseball06.model.dao.mapper.BatterReturnDtoMapper;
+import com.codesquad.baseball06.model.dao.mapper.PitcherReturnDtoMapper;
 import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
@@ -19,11 +15,11 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class PlayerDao {
 
-  private NamedParameterJdbcTemplate jdbcTemplate;
+  private static final Logger log = LoggerFactory.getLogger(PlayerDao.class);
+
   private final PitcherReturnDtoMapper pitcherReturnDtoMapper = new PitcherReturnDtoMapper();
   private final BatterReturnDtoMapper batterReturnDtoMapper = new BatterReturnDtoMapper();
-
-  public PlayerDao() {}
+  private NamedParameterJdbcTemplate jdbcTemplate;
 
   public PlayerDao(DataSource dataSource) {
     this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
@@ -31,11 +27,15 @@ public class PlayerDao {
 
   public PitcherReturnDto findPitcherById(Long id) {
 
-//    String sql = "SELECT p.id, p.team_id, p.type, p.name, p.batting_average FROM player p WHERE p.id = 1";
-    String sql = "SELECT * FROM player";
+    String sql = "SELECT p.id, p.team_id, p.type, p.name, p.batting_average FROM player p WHERE p.id = 1";
+//    String sql = "SELECT * FROM player";
     SqlParameterSource namedParameters = new MapSqlParameterSource()
         .addValue("id", id);
-    return jdbcTemplate.queryForObject(sql, namedParameters, pitcherReturnDtoMapper);
+
+    PitcherReturnDto pitcherReturnDto = jdbcTemplate
+        .queryForObject(sql, namedParameters, pitcherReturnDtoMapper);
+
+    return pitcherReturnDto;
   }
 
   public BatterReturnDto findBatterById(Long id) {
@@ -56,6 +56,7 @@ public class PlayerDao {
         .addValue("type", type)
         .addValue("name", name)
         .addValue("batting_average", batting_average);
+
     jdbcTemplate.update(sql, namedParameters);
   }
 
@@ -68,6 +69,7 @@ public class PlayerDao {
         .addValue("team_id", team_id)
         .addValue("type", type)
         .addValue("name", name);
+
     jdbcTemplate.update(sql, namedParameters);
   }
 }
